refactor: split DiffView into algorithm, scroll, and view modules
DiffView.tsx was 1642 lines mixing four separable concerns. Split along
the seams that already existed:

- diff-rows.ts       the DOM-free row model: line pairing, whitespace-
                     insensitive pairing, word-diff, and their give-up
                     budgets. This is the part with real heuristics, and
                     it is now testable without rendering.
- diff-scroll.ts     scroll targeting and viewport geometry.
- DiffOverview.tsx   the overview rail and viewport thumb.
- DiffAnnotations.tsx findings rendered as pinned rows / banner / message.

DiffView.tsx keeps the diff table itself (1642 -> 1062 lines). Public
surface is unchanged; test/diff-view.test.ts now imports the pure helpers
from their own modules instead of reaching through the component file.
