+++ src/components/DiffView.tsx
@@ -1,18 +1,20 @@
-import { diffArrays, diffLines, diffWordsWithSpace, type ChangeObject } from "diff";
 import { Fragment, type ComponentChildren } from "preact";
-import { useComputed, useSignal, type Signal } from "@preact/signals";
-import { Show } from "@preact/signals/utils";
+import { useSignal, type Signal } from "@preact/signals";
 import { useEffect, useMemo, useRef } from "preact/hooks";
-import { Badge, severityTone, statusTone } from "./Badge";
-import {
-  annotationLabel,
-  partitionFindingsByLine,
-  severityGroup,
-  type DiffFinding,
-  type SeverityGroup,
-} from "./diff-annotations";
+import { Badge, statusTone } from "./Badge";
+import { partitionFindingsByLine, type DiffFinding } from "./diff-annotations";
 import { buildDisplaySegments, GAP_EXPAND_STEP, GAP_SHOW_ALL_MAX } from "./diff-hunks";
-import { diffOverviewMarkers, displayOverviewRows, type DiffOverviewMarker } from "./diff-overview";
+import { diffOverviewMarkers, displayOverviewRows } from "./diff-overview";
+import { DiffOverview } from "./DiffOverview";
+import { AnnotationBanner, AnnotationRows, DiffMessage } from "./DiffAnnotations";
+import { buildRows, splitLines, type Row, type WordPart } from "./diff-rows";
+import {
+  initialScrollResetKey,
+  isDiffScrollTarget,
+  resetDiffScroll,
+  shouldSeekInitialDiffTarget,
+  type DiffScrollState,
+} from "./diff-scroll";
 import {
   formatLanguageFor,
   formatSourcePair,
@@ -53,375 +55,6 @@ export interface DiffViewProps {
   findings?: DiffFinding[];
 }
 
-interface Row {
-  tone: "added" | "removed" | "unchanged";
-  beforeLine: number | null;
-  afterLine: number | null;
-  text: string;
-  tokens: TokenLine | null;
-  wordParts: WordPart[] | null;
-}
-
-interface WordPart {
-  text: string;
-  tone: "added" | "removed" | "unchanged";
-}
-
-interface RowChunk {
-  tone: Row["tone"];
-  rows: Row[];
-}
-
-interface DiffOptions {
-  wordDiff?: boolean;
-  ignoreWhitespace?: boolean;
-  // Line-pairing budget, defaulting to LINE_DIFF_TIMEOUT_MS. Only tests pass it,
-  // so the give-up path below is reachable without staging a four-second diff.
-  timeoutMs?: number;
-}
-
-const INITIAL_SCROLL_TARGET_SELECTOR = "[data-diff-scroll-target='true']";
-const INITIAL_SCROLL_PADDING = 8;
-
-export interface DiffRows {
-  rows: Row[];
-  // False when line pairing was abandoned. The caller renders each source side
-  // independently and incrementally instead of manufacturing changed rows.
-  paired: boolean;
-}
-
-export function buildRows(
-  before: string,
-  after: string,
-  beforeTokens: TokenLine[] | null,
-  afterTokens: TokenLine[] | null,
-  options: DiffOptions = {},
-): DiffRows {
-  const timeoutMs = options.timeoutMs ?? LINE_DIFF_TIMEOUT_MS;
-  const result = options.ignoreWhitespace
-    ? buildRowsIgnoringWhitespace(before, after, beforeTokens, afterTokens, timeoutMs)
-    : buildRowsFromLineDiff(before, after, beforeTokens, afterTokens, timeoutMs);
-  if (options.wordDiff) applyWordDiff(result.chunks);
-  return { rows: result.chunks.flatMap((chunk) => chunk.rows), paired: result.paired };
-}
-
-interface DiffChunks {
-  chunks: RowChunk[];
-  paired: boolean;
-}
-
-export function shouldSeekInitialDiffTarget(status: string): boolean {
-  return status === "added" || status === "removed" || status === "modified";
-}
-
-export function isDiffScrollTarget(
-  status: string,
-  tone: "added" | "removed" | "unchanged",
-): boolean {
-  return shouldSeekInitialDiffTarget(status) && tone !== "unchanged";
-}
-
-// Myers diff is O(N·D), and reformatting a minified bundle turns a 1×1 line
-// comparison into a several-thousand-line one. Two rebuilds of the same source
-// keep D small and finish in tens of milliseconds; two *unrelated* bundles at
-// the 128 KiB sample cap take about four seconds of blocked main thread. This
-// bound keeps that case from freezing the tab, and `paired: false` tells the
-// reader the rows below are a whole-file replacement rather than a real pairing.
-const LINE_DIFF_TIMEOUT_MS = 1500;
-
-function buildRowsFromLineDiff(
-  before: string,
-  after: string,
-  beforeTokens: TokenLine[] | null,
-  afterTokens: TokenLine[] | null,
-  timeoutMs: number,
-): DiffChunks {
-  const parts = diffLines(before, after, { timeout: timeoutMs });
-  if (!parts) {
-    return { chunks: [], paired: false };
-  }
-  const chunks: RowChunk[] = [];
-  let beforeLine = 0;
-  let afterLine = 0;
-  for (const part of parts) {
-    const rows: Row[] = [];
-    const lines = splitLines(part.value);
-    for (const line of lines) {
-      if (part.added) {
-        afterLine += 1;
-        rows.push({
-          tone: "added",
-          beforeLine: null,
-          afterLine,
-          text: line,
-          tokens: afterTokens?.[afterLine - 1] ?? null,
-          wordParts: null,
-        });
-      } else if (part.removed) {
-        beforeLine += 1;
-        rows.push({
-          tone: "removed",
-          beforeLine,
-          afterLine: null,
-          text: line,
-          tokens: beforeTokens?.[beforeLine - 1] ?? null,
-          wordParts: null,
-        });
-      } else {
-        beforeLine += 1;
-        afterLine += 1;
-        rows.push({
-          tone: "unchanged",
-          beforeLine,
-          afterLine,
-          text: line,
-          tokens: afterTokens?.[afterLine - 1] ?? beforeTokens?.[beforeLine - 1] ?? null,
-          wordParts: null,
-        });
-      }
-    }
-    if (rows.length) {
-      chunks.push({
-        tone: part.added ? "added" : part.removed ? "removed" : "unchanged",
-        rows,
-      });
-    }
-  }
-  return { chunks, paired: true };
-}
-
-function buildRowsIgnoringWhitespace(
-  before: string,
-  after: string,
-  beforeTokens: TokenLine[] | null,
-  afterTokens: TokenLine[] | null,
-  timeoutMs: number,
-): DiffChunks {
-  const parts = diffArrays(splitLines(before), splitLines(after), {
-    comparator: linesEqualIgnoringWhitespace,
-    timeout: timeoutMs,
-  });
-  if (!parts) {
-    return { chunks: [], paired: false };
-  }
-  const chunks: RowChunk[] = [];
-  let beforeLine = 0;
-  let afterLine = 0;
-  for (const part of parts) {
-    const rows: Row[] = [];
-    for (const line of part.value) {
-      if (part.added) {
-        afterLine += 1;
-        rows.push({
-          tone: "added",
-          beforeLine: null,
-          afterLine,
-          text: line,
-          tokens: afterTokens?.[afterLine - 1] ?? null,
-          wordParts: null,
-        });
-      } else if (part.removed) {
-        beforeLine += 1;
-        rows.push({
-          tone: "removed",
-          beforeLine,
-          afterLine: null,
-          text: line,
-          tokens: beforeTokens?.[beforeLine - 1] ?? null,
-          wordParts: null,
-        });
-      } else {
-        beforeLine += 1;
-        afterLine += 1;
-        rows.push({
-          tone: "unchanged",
-          beforeLine,
-          afterLine,
-          text: line,
-          tokens: afterTokens?.[afterLine - 1] ?? beforeTokens?.[beforeLine - 1] ?? null,
-          wordParts: null,
-        });
-      }
-    }
-    if (rows.length) {
-      chunks.push({
-        tone: part.added ? "added" : part.removed ? "removed" : "unchanged",
-        rows,
-      });
-    }
-  }
-  return { chunks, paired: true };
-}
-
-function splitLines(value: string): string[] {
-  const lines = value.split("\n");
-  if (lines.length && lines[lines.length - 1] === "") lines.pop();
-  // Strip the CR of CRLF endings: shiki's tokenizer drops it, so a CRLF file
-  // would otherwise render differently plain vs highlighted (a stray \r is a
-  // segment break under whitespace-pre-wrap in some engines). Diffing still
-  // runs on the raw samples, so both sides stay consistent.
-  for (let index = 0; index < lines.length; index += 1) {
-    const line = lines[index];
-    if (line.endsWith("\r")) lines[index] = line.slice(0, -1);
-  }
-  return lines;
-}
-
-function linesEqualIgnoringWhitespace(left: string, right: string): boolean {
-  return stripWhitespace(left) === stripWhitespace(right);
-}
-
-function stripWhitespace(value: string): string {
-  return value.replace(/\s+/g, "");
-}
-
-// pairChangedRows scores every removed×added line pair with a word diff, so a
-// single huge changed block (common in bundled artifacts) turns quadratic.
-// Beyond this cell budget the word-diff decoration is skipped for that block;
-// line tones still render.
-const WORD_DIFF_MAX_PAIR_CELLS = 10_000;
-
-function applyWordDiff(chunks: RowChunk[]) {
-  for (let index = 0; index < chunks.length - 1; index += 1) {
-    const removed = chunks[index];
-    const added = chunks[index + 1];
-    if (removed.tone !== "removed" || added.tone !== "added") continue;
-    if (removed.rows.length * added.rows.length > WORD_DIFF_MAX_PAIR_CELLS) continue;
-    for (const [beforeRow, afterRow] of pairChangedRows(removed.rows, added.rows)) {
-      const parts = buildWordParts(beforeRow.text, afterRow.text);
-      beforeRow.wordParts = parts.before;
-      afterRow.wordParts = parts.after;
-    }
-  }
-}
-
-function pairChangedRows(beforeRows: Row[], afterRows: Row[]): [Row, Row][] {
-  const scores = beforeRows.map((beforeRow) =>
-    afterRows.map((afterRow) => rowPairScore(beforeRow.text, afterRow.text)),
-  );
-  const dp = Array.from({ length: beforeRows.length + 1 }, () =>
-    Array.from({ length: afterRows.length + 1 }, () => 0),
-  );
-  const choices: ("match" | "before" | "after")[][] = Array.from(
-    { length: beforeRows.length + 1 },
-    () => Array.from({ length: afterRows.length + 1 }, () => "before"),
-  );
-
-  for (let beforeIndex = 1; beforeIndex <= beforeRows.length; beforeIndex += 1) {
-    for (let afterIndex = 1; afterIndex <= afterRows.length; afterIndex += 1) {
-      const score = scores[beforeIndex - 1][afterIndex - 1];
-      const match = score > 0 ? dp[beforeIndex - 1][afterIndex - 1] + score : -Infinity;
-      const skipBefore = dp[beforeIndex - 1][afterIndex];
-      const skipAfter = dp[beforeIndex][afterIndex - 1];
-      if (match > skipBefore && match > skipAfter) {
-        dp[beforeIndex][afterIndex] = match;
-        choices[beforeIndex][afterIndex] = "match";
-      } else if (skipBefore >= skipAfter) {
-        dp[beforeIndex][afterIndex] = skipBefore;
-        choices[beforeIndex][afterIndex] = "before";
-      } else {
-        dp[beforeIndex][afterIndex] = skipAfter;
-        choices[beforeIndex][afterIndex] = "after";
-      }
-    }
-  }
-
-  const pairs: [Row, Row][] = [];
-  let beforeIndex = beforeRows.length;
-  let afterIndex = afterRows.length;
-  while (beforeIndex > 0 && afterIndex > 0) {
-    const choice = choices[beforeIndex][afterIndex];
-    if (choice === "match") {
-      pairs.push([beforeRows[beforeIndex - 1], afterRows[afterIndex - 1]]);
-      beforeIndex -= 1;
-      afterIndex -= 1;
-    } else if (choice === "before") {
-      beforeIndex -= 1;
-    } else {
-      afterIndex -= 1;
-    }
-  }
-  return pairs.reverse();
-}
-
-const linePairScoreFloor = 0.08;
-
-function rowPairScore(before: string, after: string): number {
-  if (!compatibleLineKinds(before, after)) return 0;
-  const beforeText = before.trim();
-  const afterText = after.trim();
-  if (!beforeText || !afterText) return 0;
-
-  let unchangedLength = 0;
-  for (const part of diffWordsWithSpace(beforeText, afterText) as ChangeObject<string>[]) {
-    if (!part.added && !part.removed) unchangedLength += part.value.length;
-  }
-  const score = unchangedLength / Math.max(beforeText.length, afterText.length);
-  return score >= linePairScoreFloor ? score : 0;
-}
-
-function compatibleLineKinds(before: string, after: string): boolean {
-  return lineKind(before) === lineKind(after);
-}
-
-function lineKind(text: string): "blank" | "comment" | "code" {
-  const trimmed = text.trim();
-  if (!trimmed) return "blank";
-  if (
-    trimmed.startsWith("//") ||
-    trimmed.startsWith("/*") ||
-    trimmed.startsWith("*") ||
-    trimmed.startsWith("#")
-  ) {
-    return "comment";
-  }
-  return "code";
-}
-
-function buildWordParts(before: string, after: string): { before: WordPart[]; after: WordPart[] } {
-  const beforeParts: WordPart[] = [];
-  const afterParts: WordPart[] = [];
-  for (const part of diffWordsWithSpace(before, after) as ChangeObject<string>[]) {
-    const tone = partTone(part);
-    if (!part.added) appendWordParts(beforeParts, part.value, tone);
-    if (!part.removed) appendWordParts(afterParts, part.value, tone);
-  }
-  return { before: beforeParts, after: afterParts };
-}
-
-const wordDiffWordPattern = /[\p{L}\p{N}_$]+/gu;
-
-function appendWordParts(parts: WordPart[], text: string, tone: WordPart["tone"]) {
-  if (tone === "unchanged") {
-    appendWordPart(parts, text, tone);
-    return;
-  }
-  let offset = 0;
-  for (const match of text.matchAll(wordDiffWordPattern)) {
-    const index = match.index;
-    if (index > offset) appendWordPart(parts, text.slice(offset, index), "unchanged");
-    appendWordPart(parts, match[0], tone);
-    offset = index + match[0].length;
-  }
-  if (offset < text.length) appendWordPart(parts, text.slice(offset), "unchanged");
-}
-
-function appendWordPart(parts: WordPart[], text: string, tone: WordPart["tone"]) {
-  if (!text) return;
-  const previous = parts[parts.length - 1];
-  if (previous?.tone === tone) {
-    previous.text += text;
-    return;
-  }
-  parts.push({ text, tone });
-}
-
-function partTone(part: ChangeObject<string>): WordPart["tone"] {
-  if (part.added) return "added";
-  if (part.removed) return "removed";
-  return "unchanged";
-}
-
 // Tokenize an entire side once, memoized on the text/language/ready signal.
 // Sides beyond the highlight cap stay plain text: tokenizing them would block
 // the main thread for seconds (see HIGHLIGHT_MAX_LINES).
@@ -438,34 +71,6 @@ function useLineTokens(text: string, lang: string | undefined): TokenLine[] | nu
     () => (lang && ready && text && canHighlight(text) ? tokenizeLines(text, lang) : null),
     [text, lang, ready],
   );
-}
-
-// Keyed on the file identity and content only — deliberately not on findings.
-// Findings arriving after the diff is on screen must not yank the scroll back
-// to the first change (or reset the single-sided render window) mid-read.
-// Exported for tests.
-export function initialScrollResetKey(
-  path: string,
-  status: string,
-  beforeSample: string,
-  afterSample: string,
-): string {
-  return [
-    path,
-    status,
-    beforeSample.length,
-    afterSample.length,
-    beforeSample.slice(0, 64),
-    afterSample.slice(0, 64),
-  ].join("\0");
-}
-
-// Scroll geometry mirrored into a signal so the overview thumb can track the
-// viewport without rerendering the diff table on every scroll frame.
-interface DiffScrollState {
-  top: number;
-  viewport: number;
-  content: number;
 }
 
 function DiffScrollViewport({
@@ -533,19 +138,6 @@ function DiffScrollViewport({
       {children}
     </div>
   );
-}
-
-function resetDiffScroll(container: HTMLElement) {
-  const target = container.querySelector<HTMLElement>(INITIAL_SCROLL_TARGET_SELECTOR);
-  if (!target) {
-    container.scrollTop = 0;
-    return;
-  }
-  const targetTop =
-    target.getBoundingClientRect().top -
-    container.getBoundingClientRect().top +
-    container.scrollTop;
-  container.scrollTop = Math.max(0, targetTop - INITIAL_SCROLL_PADDING);
 }
 
 function LineContent({
@@ -1460,179 +1052,7 @@ function SingleSidedView({
   );
 }
 
-function DiffOverview({
-  markers,
-  scrollState,
-}: {
-  markers: DiffOverviewMarker[];
-  scrollState: Signal<DiffScrollState | null>;
-}) {
-  // A pane short enough to show everything has nothing to navigate to, and the
-  // rail next to it reads as decoration rather than signal. Gated through Show
-  // so scroll frames flip one boundary instead of rerendering every marker.
-  const overflowing = useComputed(() => {
-    const state = scrollState.value;
-    return state !== null && state.content > state.viewport + 1;
-  });
-  if (!markers.length) return null;
-  return (
-    <Show when={overflowing}>
-      {() => (
-        <div
-          aria-hidden="true"
-          class="pointer-events-none absolute right-1 top-2 bottom-2 w-[6px] rounded-full border border-border bg-surface/80 shadow-sm"
-        >
-          {markers.map((marker) => (
-            <span
-              key={marker.key}
-              class={cn(
-                "absolute left-[-1px] right-[-1px] rounded-full",
-                markerClass(marker),
-                marker.kind === "finding" ? "z-10 ring-1 ring-surface" : "opacity-75",
-              )}
-              style={{
-                top: `${marker.topPercent}%`,
-                height: `${marker.heightPercent}%`,
-              }}
-            />
-          ))}
-          <DiffOverviewThumb scrollState={scrollState} />
-        </div>
-      )}
-    </Show>
-  );
-}
-
-// The viewport's position over the strip, in the same neutral ink used for
-// structure (never a severity hue — color = signal). Reads the scroll signal
-// in its own component so scrolling rerenders only this span, never the diff
-// table.
-function DiffOverviewThumb({ scrollState }: { scrollState: Signal<DiffScrollState | null> }) {
-  const state = scrollState.value;
-  if (!state || state.content <= state.viewport) return null;
-  return (
-    <span
-      class="absolute left-[-1px] right-[-1px] z-20 rounded-full bg-ink/15"
-      style={{
-        top: `${(state.top / state.content) * 100}%`,
-        height: `${(state.viewport / state.content) * 100}%`,
-      }}
-    />
-  );
-}
-
-function markerClass(marker: DiffOverviewMarker): string {
-  if (marker.tone === "added" || marker.tone === "ok") return "bg-ok";
-  if (marker.tone === "removed" || marker.tone === "danger") return "bg-danger";
-  if (marker.tone === "warn") return "bg-warn";
-  return "bg-info";
-}
-
-// Severity-tinted fills and left bars for a pinned finding. The fill uses the
-// soft severity token at reduced opacity so it reads as a callout over the
-// green/red row backgrounds; the bar uses the saturated token (shapes, per
-// docs/design.md "color = signal"). Both are static class strings so Tailwind keeps
 // them.
-const ANNOTATION_FILL: Record<SeverityGroup, string> = {
-  danger: "bg-danger-soft/60",
-  warn: "bg-warn-soft/60",
-  info: "bg-info-soft/60",
-  ok: "bg-ok-soft/60",
-};
-
-const ANNOTATION_BAR: Record<SeverityGroup, string> = {
-  danger: "border-danger",
-  warn: "border-warn",
-  info: "border-info",
-  ok: "border-ok",
-};
-
-// The body of a pinned finding: severity Badge + mono `ruleId · line N` caption,
-// the reason, and (when present) the triggering evidence in mono. Mirrors the
-// landing page's review-preview annotation so what we advertise matches the app.
-function FindingAnnotationBody({ finding }: { finding: DiffFinding }) {
-  const group = severityGroup(finding.severity);
-  const label = annotationLabel(finding);
-  return (
-    <div
-      class={cn(
-        // font-sans: the callout sits inside the mono diff table, but the reason
-        // is body copy and must not inherit Geist Mono (label/evidence set their
-        // own mono explicitly).
-        "border-l-2 px-3 py-2.5 flex flex-col gap-1.5 font-sans",
-        ANNOTATION_FILL[group],
-        ANNOTATION_BAR[group],
-      )}
-    >
-      <div class="flex flex-wrap items-center gap-2">
-        <Badge tone={severityTone(finding.severity)}>{finding.severity}</Badge>
-        {label ? (
-          <span class="font-mono text-[10px] uppercase tracking-[0.1em] text-ink-subtle">
-            {label}
-          </span>
-        ) : null}
-      </div>
-      <p class="m-0 text-[13px] leading-[1.55] text-ink whitespace-normal">{finding.reason}</p>
-      {finding.evidence ? (
-        <code class="font-mono text-[11px] leading-[1.5] text-ink-muted break-words whitespace-pre-wrap">
-          {finding.evidence}
-        </code>
-      ) : null}
-    </div>
-  );
-}
-
-// Findings pinned beneath their diff line, rendered as full-width table rows.
-function AnnotationRows({
-  findings,
-  colSpan,
-  scrollTarget = false,
-}: {
-  findings: DiffFinding[];
-  colSpan: number;
-  scrollTarget?: boolean;
-}) {
-  return (
-    <>
-      {findings.map((finding) => (
-        <tr key={finding.id} data-diff-scroll-target={scrollTarget ? "true" : undefined}>
-          <td colSpan={colSpan} class="p-0">
-            <FindingAnnotationBody finding={finding} />
-          </td>
-        </tr>
-      ))}
-    </>
-  );
-}
-
-// Findings that can't be pinned (no line, or a line past a truncated sample),
-// shown between the header strip and the scroll region so they're never hidden.
-function AnnotationBanner({ findings }: { findings: DiffFinding[] }) {
-  return (
-    <div class="flex flex-col divide-y divide-border border-b border-border">
-      {findings.map((finding) => (
-        <FindingAnnotationBody key={finding.id} finding={finding} />
-      ))}
-    </div>
-  );
-}
-
-// A non-table fallback (binary / no-sample) that still surfaces any findings as
-// standalone callouts above the explanatory line, so a missing diff body never
-// drops a deterministic signal.
-function DiffMessage({ children, findings }: { children: string; findings: DiffFinding[] }) {
-  if (!findings.length) return <Muted class="text-[13px]">{children}</Muted>;
-  return (
-    <div class="flex flex-col gap-3">
-      <div class="border border-border rounded-md overflow-hidden flex flex-col divide-y divide-border">
-        {findings.map((finding) => (
-          <FindingAnnotationBody key={finding.id} finding={finding} />
-        ))}
-      </div>
-      <Muted class="text-[13px]">{children}</Muted>
-    </div>
-  );
-}
 
 function formatSize(value: number | null): string {
   if (value === null || value === undefined) return "—";
